Create SidebarTemplate for reusability
