+++ src/routes.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Routes as ReactRouterRoutes, Route } from 'react-router-
 
 import { Calendar } from './pages/calendar';
 import { Home } from './pages/home';
-import { Scheduler } from './pages/scheduler/scheduler';
+import { Scheduler } from './pages/scheduler';
 
 // TODO: use nested routes but it doesn't work right now
